Add files via upload
This is just a test run to see if i can get output with my files. I had to convert the fasta files to .txt, and I used the AWD_LSTM language model as I am still learning how to use the BERT architecture with fastai. I will update as I make progress.